implemented socke.io for aytomatically products stock updates
